Added Base Files
Added the initial base files.

In this case it was upon addition of audio tech using SDL_Mixer.

Changes need to be made in order to put this into a separate header and
class files. Code needs to be tidied up, sprite work needs to be added,
and the game actually needs to be made.