Menu changes
The game now, for all intents and purposes, has a main menu that can be
navigated using the 'w' and 's' keys to move between the two options.
The menu doesn't lead anywhere yet. Still needs an actual game and
options screen.

An additional spritesheet has been added to the assets folder, in time
this will likely take the place of the enemy sprites in the game.