Menu Changes and File Additions
The menu has been changed to now allow the user to press the 'Return'
key to enter the different menu options. This then changes a variable,
telling the renderer which scene the game is in. These scenes currently
output messages saying that they're being rendered, this happens a lot
so that's why the console starts printing loads of messages.

The assets folder has also had some files added to it, in particular a
small blue button (will be the seeds), a tall yellow button (to be the
eggs), a middle section of a ladder and a mid section of stone to be the
floors