Changes
A royalty free music file has been added to the assets folder and has
replaced the previous placeholder. This music now plays upon the player
entering the game.

For sprites, a bunch of code has been added to that class to draw both
players and an enemy. The level is slightly more populated with floors.
Collision detection has also been added for both player1 and player2, to
ensure that they don't walk off the edge of the screen. However, player1
does seem to have a tendency to moonwalk, which needs to be fixed.

The players need to have collision detection for the different floors,
ladders and items. This then needs to impact the scoring system.

Speaking of which, a scoring system needs to be implemented and
subsequently needs to be displayed in game. This needs to have a scoring
system and a timer, preferably.

Player 1 and Player 2 are controllable on the x-axis, with a/d and
left/right respectively.